+++ src/utils/Profile.ts
@@ -1,6 +1,6 @@
 import { EOL } from 'node:os';
 import chalk from 'chalk';
-import { uuids } from '../assigned numbers/16-bit UUID Numbers.js';
+import { uuids } from '../../assigned-numbers/16-bit UUID Numbers.js';
 import { Profile, Service, Characteristic, IncludedService, Descriptor } from '../gatt/GattDirectory.js';
 
 export const uuidInfo = (uuid: string): { type: string; for: string; } | undefined => uuids.entries[uuid];