+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.8] - 2023-12-03
+
+- rename dist -> lib in release files structure
+- deliver assigned numbers in lib/assigned-numbers
+- update outdated
+- process advertisement report appearance field
+
 ## [1.0.7] - 2023-11-19
 
 - Update package.json